mcasp: PRSDK-7455: Example fix on j721e/c66x
Board configuration flag to unlock MMR is not functional on c66x because
Board library does not comprehend c66x memory map on this device.

Previously, this issue was masked because unlock MMR function did not
return error. Recently this function was updated, so MMR unlock returns
error and causes subsequent board initialization failures on c66x. This
causes CPU exception when trying to printf via UART but UART
initialization was not run due to Board failures.

Disabling MMR unlock since it is not needed for this application at this
point.

Signed-off-by: Stephen Molfetta <sjmolfetta@ti.com>