prsdk-7487: icssg f/w: use the correct spinlocks to report EOF to RTU
In half duplex mode we cannot report EOF in TX_FIFO handler because packet
may be resent. Instead the report has to be done in the TX_EOF handler.
The current implementation by mistake uses the wrong spinlocks for the
report. Thus it uses spinlock E for the DMA0 and spinlock P for the DMA1,
which is wrong. This commit fixes this bug.

Signed-off-by: Vitaly Andrianov <vitalya@ti.com>