PRSDK-7315:Patch ELF binary for out2rprc tool
This script file patches an elf binary to force the exidx section to be set to type "PROGBITS".
This makes the out2rprc tool fill in the exidx section in the converted image.
This is used in case CPP exception handling is required in a ARM Coretex-A15 application.
The patching is performed in place.