prsdk-7627: emac: add set port state to rtu code
This commit add code to handle set port state mgr command.
It store the command in the dedicated smem location and tell PRU about
that. Then it waits pru finishes the command execution.
Once it is done, it updates own speed_f and sends the response to the host.

This command doesn't handle the set_port_state/disable_port case.
Also the implementation doesn't check validity of the command parameters.
Driver must not set invalid parameters.

Signed-off-by: Vitaly Andrianov <vitalya@ti.com>