PRSDK-7681: IPC: Fix for C7x random crash in ISR
There was random crash in vision_app on C7x in
release mode. When crash occurred, the PC was close
to Mailbox ISR.

C7x stack to be 0x20000 aligned for proper
unwinding of nested interrupt.
The issue was reproduced in Vision_app but all
the c7x application must have stack as 0x20000 aligned

Signed-off-by: Santosh Jha <sjha@ti.com>